更新 main.blocks, main.ts

+++ main.ts
@@ -8,49 +8,57 @@ input.onButtonPressed(Button.A, function () {
 })
 // 1、2、3、4、5P五連發
 input.onGesture(Gesture.Shake, function () {
-    _1p1 = game.createSprite(2, 4)
-    _2p1 = game.createSprite(0, 4)
-    _3p = game.createSprite(4, 4)
-    _4p = game.createSprite(1, 4)
-    _5p = game.createSprite(3, 4)
-    bullet = game.createSprite(_1p1.get(LedSpriteProperty.X), _1p1.get(LedSpriteProperty.Y))
-    bullet1 = game.createSprite(_2p1.get(LedSpriteProperty.X), _2p1.get(LedSpriteProperty.Y))
-    bullet2 = game.createSprite(_3p.get(LedSpriteProperty.X), _3p.get(LedSpriteProperty.Y))
-    bullet3 = game.createSprite(_4p.get(LedSpriteProperty.X), _4p.get(LedSpriteProperty.Y))
-    bullet4 = game.createSprite(_5p.get(LedSpriteProperty.X), _5p.get(LedSpriteProperty.Y))
-    for (let index = 0; index < 5; index++) {
-        for (let index = 0; index < 4; index++) {
-            bullet.change(LedSpriteProperty.Y, -1)
-            bullet1.change(LedSpriteProperty.Y, -1)
-            bullet2.change(LedSpriteProperty.Y, -1)
-            bullet3.change(LedSpriteProperty.Y, -1)
-            bullet4.change(LedSpriteProperty.Y, -1)
-            basic.pause(100)
+    if (game.isGameOver()) {
+    	
+    } else {
+        _1p1 = game.createSprite(2, 4)
+        _2p1 = game.createSprite(0, 4)
+        _3p = game.createSprite(4, 4)
+        _4p = game.createSprite(1, 4)
+        _5p = game.createSprite(3, 4)
+        bullet = game.createSprite(_1p1.get(LedSpriteProperty.X), _1p1.get(LedSpriteProperty.Y))
+        bullet1 = game.createSprite(_2p1.get(LedSpriteProperty.X), _2p1.get(LedSpriteProperty.Y))
+        bullet2 = game.createSprite(_3p.get(LedSpriteProperty.X), _3p.get(LedSpriteProperty.Y))
+        bullet3 = game.createSprite(_4p.get(LedSpriteProperty.X), _4p.get(LedSpriteProperty.Y))
+        bullet4 = game.createSprite(_5p.get(LedSpriteProperty.X), _5p.get(LedSpriteProperty.Y))
+        for (let index = 0; index < 5; index++) {
+            for (let index = 0; index < 4; index++) {
+                bullet.change(LedSpriteProperty.Y, -1)
+                bullet1.change(LedSpriteProperty.Y, -1)
+                bullet2.change(LedSpriteProperty.Y, -1)
+                bullet3.change(LedSpriteProperty.Y, -1)
+                bullet4.change(LedSpriteProperty.Y, -1)
+                basic.pause(100)
+            }
         }
+        bullet.delete()
+        bullet1.delete()
+        bullet2.delete()
+        bullet3.delete()
+        bullet4.delete()
+        basic.pause(100)
+        _1p1.delete()
+        _2p1.delete()
+        _3p.delete()
+        _4p.delete()
+        _5p.delete()
     }
-    bullet.delete()
-    bullet1.delete()
-    bullet2.delete()
-    bullet3.delete()
-    bullet4.delete()
-    basic.pause(100)
-    _1p1.delete()
-    _2p1.delete()
-    _3p.delete()
-    _4p.delete()
-    _5p.delete()
 })
 // 1P+2P攻
 input.onButtonPressed(Button.AB, function () {
-    bullet = game.createSprite(_1p.get(LedSpriteProperty.X), _1p.get(LedSpriteProperty.Y))
-    bullet1 = game.createSprite(_2p.get(LedSpriteProperty.X), _2p.get(LedSpriteProperty.Y))
-    for (let index = 0; index < 4; index++) {
-        bullet.change(LedSpriteProperty.Y, -1)
-        bullet1.change(LedSpriteProperty.Y, -1)
-        basic.pause(100)
+    if (game.isGameOver()) {
+    	
+    } else {
+        bullet = game.createSprite(_1p.get(LedSpriteProperty.X), _1p.get(LedSpriteProperty.Y))
+        bullet1 = game.createSprite(_2p.get(LedSpriteProperty.X), _2p.get(LedSpriteProperty.Y))
+        for (let index = 0; index < 4; index++) {
+            bullet.change(LedSpriteProperty.Y, -1)
+            bullet1.change(LedSpriteProperty.Y, -1)
+            basic.pause(100)
+        }
+        bullet.delete()
+        bullet1.delete()
     }
-    bullet.delete()
-    bullet1.delete()
 })
 // 1P右
 input.onButtonPressed(Button.B, function () {
@@ -91,53 +99,57 @@ basic.forever(function () {
 })
 // 分數計算
 basic.forever(function () {
-    if (bullet) {
-        if (bullet.isTouching(airplane)) {
-            game.addScore(1e+284)
-            airplane.set(LedSpriteProperty.X, 0)
-            airplane.set(LedSpriteProperty.Y, 0)
+    if (game.isGameOver()) {
+    	
+    } else {
+        if (bullet) {
+            if (bullet.isTouching(airplane)) {
+                game.addScore(1e+284)
+                airplane.set(LedSpriteProperty.X, 0)
+                airplane.set(LedSpriteProperty.Y, 0)
+            }
         }
-    }
-    if (bullet1) {
-        if (bullet1.isTouching(airplane)) {
-            game.addScore(1e+284)
-            airplane.set(LedSpriteProperty.X, 0)
-            airplane.set(LedSpriteProperty.Y, 0)
+        if (bullet1) {
+            if (bullet1.isTouching(airplane)) {
+                game.addScore(1e+284)
+                airplane.set(LedSpriteProperty.X, 0)
+                airplane.set(LedSpriteProperty.Y, 0)
+            }
         }
-    }
-    if (bullet2) {
-        if (bullet2.isTouching(airplane)) {
-            game.addScore(1e+284)
-            airplane.set(LedSpriteProperty.X, 0)
-            airplane.set(LedSpriteProperty.Y, 0)
+        if (bullet2) {
+            if (bullet2.isTouching(airplane)) {
+                game.addScore(1e+284)
+                airplane.set(LedSpriteProperty.X, 0)
+                airplane.set(LedSpriteProperty.Y, 0)
+            }
         }
-    }
-    if (bullet3) {
-        if (bullet3.isTouching(airplane)) {
-            game.addScore(1e+284)
-            airplane.set(LedSpriteProperty.X, 0)
-            airplane.set(LedSpriteProperty.Y, 0)
+        if (bullet3) {
+            if (bullet3.isTouching(airplane)) {
+                game.addScore(1e+284)
+                airplane.set(LedSpriteProperty.X, 0)
+                airplane.set(LedSpriteProperty.Y, 0)
+            }
         }
-    }
-    if (bullet4) {
-        if (bullet4.isTouching(airplane)) {
-            game.addScore(1e+284)
-            airplane.set(LedSpriteProperty.X, 0)
-            airplane.set(LedSpriteProperty.Y, 0)
+        if (bullet4) {
+            if (bullet4.isTouching(airplane)) {
+                game.addScore(1e+284)
+                airplane.set(LedSpriteProperty.X, 0)
+                airplane.set(LedSpriteProperty.Y, 0)
+            }
         }
-    }
-    if (missile1) {
-        if (missile1.isTouching(airplane)) {
-            game.addScore(1e+284)
-            airplane.set(LedSpriteProperty.X, 0)
-            airplane.set(LedSpriteProperty.Y, 0)
+        if (missile1) {
+            if (missile1.isTouching(airplane)) {
+                game.addScore(1e+284)
+                airplane.set(LedSpriteProperty.X, 0)
+                airplane.set(LedSpriteProperty.Y, 0)
+            }
         }
-    }
-    if (missile2) {
-        if (missile2.isTouching(missile2)) {
-            game.addScore(1e+284)
-            missile2.set(LedSpriteProperty.X, 0)
-            missile2.set(LedSpriteProperty.Y, 0)
+        if (missile2) {
+            if (missile2.isTouching(missile2)) {
+                game.addScore(1e+284)
+                missile2.set(LedSpriteProperty.X, 0)
+                missile2.set(LedSpriteProperty.Y, 0)
+            }
         }
     }
 })
@@ -149,35 +161,43 @@ basic.forever(function () {
 })
 // 導彈(1)
 basic.forever(function () {
-    basic.pause(randint(5000, 10000))
-    missile1 = game.createSprite(2, 4)
-    missile1.turn(Direction.Right, 45)
-    basic.pause(100)
-    missile1.move(-1)
-    basic.pause(100)
-    missile1.move(-1)
-    basic.pause(100)
-    missile1.turn(Direction.Left, 90)
-    missile1.move(1)
-    basic.pause(100)
-    missile1.move(1)
-    basic.pause(100)
-    missile1.delete()
+    if (game.isGameOver()) {
+    	
+    } else {
+        basic.pause(randint(5000, 10000))
+        missile1 = game.createSprite(2, 4)
+        missile1.turn(Direction.Right, 45)
+        basic.pause(100)
+        missile1.move(-1)
+        basic.pause(100)
+        missile1.move(-1)
+        basic.pause(100)
+        missile1.turn(Direction.Left, 90)
+        missile1.move(1)
+        basic.pause(100)
+        missile1.move(1)
+        basic.pause(100)
+        missile1.delete()
+    }
 })
 // 導彈(2)
 basic.forever(function () {
-    basic.pause(randint(5000, 10000))
-    missile2 = game.createSprite(2, 4)
-    missile2.turn(Direction.Left, 45)
-    basic.pause(100)
-    missile2.move(1)
-    basic.pause(100)
-    missile2.move(1)
-    basic.pause(100)
-    missile2.turn(Direction.Right, 90)
-    missile2.move(-1)
-    basic.pause(100)
-    missile2.move(-1)
-    basic.pause(100)
-    missile2.delete()
+    if (game.isGameOver()) {
+    	
+    } else {
+        basic.pause(randint(5000, 10000))
+        missile2 = game.createSprite(2, 4)
+        missile2.turn(Direction.Left, 45)
+        basic.pause(100)
+        missile2.move(1)
+        basic.pause(100)
+        missile2.move(1)
+        basic.pause(100)
+        missile2.turn(Direction.Right, 90)
+        missile2.move(-1)
+        basic.pause(100)
+        missile2.move(-1)
+        basic.pause(100)
+        missile2.delete()
+    }
 })
